make clean and pretty looks and code

+++ run.py
@@ -2,13 +2,65 @@ from synth import main as syn_main
 from web_ch import main as get_main 
 from player import main as play_main 
 from lib.pytui.pytui import Tui
+from stats import get_chapters_left
 
-import sys
 import time
 import signal
 import os
+import sys
+os.makedirs("output/.logging", exist_ok=True)
+sys.stderr = open('output/.logging/run.log', 'w')
 from multiprocessing import Process, active_children
 from playa_utils import get_var
+
+def stats_thread(tui=Tui()):
+    while True:
+        tui.buffered  = True 
+        tui.place_text("Chapters ready for reading:",row = 0, height=1)
+        x = get_chapters_left()
+        l = len(str(max(x, key=lambda item: item[0])[0]))
+        for i, (c,_,_,dir) in enumerate(x):
+            tui.place_text(f"{c:>{l}}: {dir}", col = 2, row=i+1, height=1)
+        tui.flush()
+        time.sleep(10)
+
+
+def main():
+    do = "gspx"
+    if len(sys.argv) > 1:
+        do = sys.argv[1]
+    ps = []
+    tui=Tui()
+    tui.clear()
+    input_pos = (1,1)
+    bg = (0,0,64)
+    
+
+    if "g" in do:
+        get_tui = Tui(col_offset=119,row_offset=2, max_width=40,max_height = 4, default_cursor_pos=input_pos,border=u"\u2588",bg_colour=bg) 
+        get_p = Process(target=get_main, args=(get_tui,))
+        get_p.start()
+        ps += [get_p]
+
+    if "s" in do:
+        syn_tui = Tui(col_offset=119,row_offset=5,max_width=40, max_height=19, default_cursor_pos=input_pos,border=u"\u2588",bg_colour=bg) 
+        syn_p = Process(target=syn_main, args=(syn_tui,))
+        syn_p.start()
+        ps += [syn_p]
+
+    if "x" in do:
+        stat_tui = Tui(col_offset=119,row_offset=23,max_width=40, max_height=22, default_cursor_pos=input_pos,border=u"\u2588",bg_colour=bg) 
+        stat_p = Process(target=stats_thread, args=(stat_tui,))
+        stat_p.start()
+        ps += [stat_p]
+
+    if "p" in do:
+        play_tui = Tui(col_offset=0,row_offset=2,default_cursor_pos=input_pos,max_width=120, max_height=43,border=u"\u2588",bg_colour=bg) 
+        play_main(play_tui)
+        for p in ps:
+            p.terminate()
+        stop(signal.SIGTERM)
+
 def signal_handler(sig, frame):
     fn = get_var('stop')
     if fn:
@@ -17,69 +69,22 @@ def signal_handler(sig, frame):
 signal.signal(signal.SIGINT, signal_handler)
 signal.signal(signal.SIGTERM, signal_handler)
 def stop(sig):
-    syn_tui = get_var('syn_tui')
-    if syn_tui:
-        syn_tui.buffered  = True 
-        syn_tui.clear_box(row=2+lines_offset)
-        syn_tui.place_text("Synth interrupted",row = 2+lines_offset, height=1)
-        syn_tui.place_text("Chapters ready for reading:",row = 3+lines_offset, height=1)
-        x = get_chapters_left()
-        l = len(str(max(x, key=lambda item: item[0])[0]))
-        for i, (c,_,_,dir) in enumerate(x):
-            syn_tui.place_text(f"{c:>{l}}: {dir}", col = 2, row=lines_offset+4+i, height=1)
-        syn_tui.flush()
-    for child in [get_var('ui_p'), get_var('player_p'), get_var('get_p'), get_var('syn_p')]:
-        if child:
-            os.kill(child.pid, sig)
     driver = get_var('driver')
-    get_tui = get_var('get_tui')
     if driver:
         driver.quit()
-    if get_tui:
-        get_tui.place_text("Web getter Interrupted", height=1)
+    for child in [get_var('ui_p'), get_var('player_p'), get_var('get_p'), get_var('syn_p'), get_var('stat_p')]:
+        if child:
+            os.kill(child.pid, sig)
     for child in active_children():
-        os.kill(child.pid, sig)
+        if child.pid:
+            os.kill(child.pid, sig)
     time.sleep(0.1)
     for child in active_children():
-        os.kill(child.pid, signal.SIGKILL)
+        if child.pid:
+            os.kill(child.pid, signal.SIGKILL)
     if __name__ == '__main__':
-        tui=Tui(hide_cursor=False)
+        Tui(hide_cursor=False)
     exit(0)
-
-def main():
-    do = "gsp"
-    if len(sys.argv) > 1:
-        do = sys.argv[1]
-    ps = []
-    tui=Tui()
-    tui.clear()
-    input_pos = (0,30)
-    all_col_offset = 0 
-    all_row_offset = 0 
-    if "g" in do:
-        get_tui = Tui(col_offset=all_col_offset+(70 if "p" in do else 0),row_offset=all_row_offset, default_cursor_pos=input_pos) 
-        if "s" in do: 
-            get_tui.max_height = 2
-        get_p = Process(target=get_main, args=(get_tui,))
-        get_p.start()
-        ps += [get_p]
-
-    if "s" in do:
-        syn_tui = Tui(col_offset=all_col_offset+(70 if "p" in do else 0),row_offset=all_row_offset+(3 if "g" in do else 0), default_cursor_pos=input_pos) 
-        syn_p = Process(target=syn_main, args=(syn_tui,))
-        syn_p.start()
-        ps += [syn_p]
-
-    if "p" in do:
-        play_tui = Tui(col_offset=all_col_offset,row_offset=all_row_offset,default_cursor_pos=input_pos) 
-        if "s" in do or "g" in do:
-            play_tui.max_width = 60
-        play_main(play_tui)
-        stop(signal.SIGTERM)
-
-
-    for child in multiprocessing.active_children():
-        child.wait()
 
 
 if __name__ == "__main__":
